+++ tests/test_pool.py
@@ -16,7 +16,7 @@ async def test_pool_cursor(pool):
     async with pool.acquire() as conn:
         async with conn.cursor() as cursor:
             await cursor.execute("SELECT 1")
-            ret = cursor.fetchone()
+            ret = await cursor.fetchone()
             assert ret == (1,)